refactor(idea_service): extract list_ideas + get_idea to idea_read (#163)
  before: 1981
  after:  1895 (-86)
